pyproject.toml
[project]
name = "job-matching-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elasticsearch>=9.1.1",
    "fastapi>=0.117.1",
    "openai>=1.108.1",
    "pandas>=2.3.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.37.0",
]
